pyproject.toml
[tool.poetry]
name = "sarawan"
version = "0.1.0"
description = ""
authors = ["UltimaKIND <minion9@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.4"
djangorestframework = "^3.15.2"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.10"
djangorestframework-simplejwt = "^5.3.1"
pillow = "^11.0.0"
pytest-django = "^4.9.0"
python-slugify = {extras = ["unidecode"], version = "^8.0.4"}
black = "^24.10.0"
isort = "^5.13.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
